% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{add_upacc}
\alias{add_upacc}
\title{Add user and producer accuracies to a confusion matrix.}
\usage{
add_upacc(conmat)
}
\arguments{
\item{conmat}{A confusion matrix or a list.}
}
\value{
The conmat with an extra row and column, or a list.
}
\description{
Add user and produser accuracies to a confusion matrix.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
