% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{asses_accuracy_area}
\alias{asses_accuracy_area}
\title{Asses accuracy and estimate area according to Olofsson}
\usage{
asses_accuracy_area(error_matrix, class_areas)
}
\arguments{
\item{error_matrix}{A matrix given in sample counts. Columns represent the reference data and rows the results of the classification}

\item{class_areas}{A vector of the total area of each class on the map}
}
\value{
A list of lists: The error_matrix, the class_areas, confidence interval (confint95, a list of two numerics) and the accuracy (accuracy, a list of three numerics: overall, user, and producer)
}
\description{
Compute the accuracy normalized by the area. Note that, these computations don't work on clustered sampling because the equations are different.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
