% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{asses_accuracy_simple}
\alias{asses_accuracy_simple}
\title{Asses accuracy}
\usage{
asses_accuracy_simple(error_matrix)
}
\arguments{
\item{error_matrix}{A matrix given in sample counts. Columns represent the reference data and rows the results of the classification}
}
\value{
A list of numeric vectors: The overall, user and producer accuracies
}
\description{
Compute the overall, user, and producer accuracies.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
