% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brick_description}
\alias{brick_description}
\title{Description of the Bricks used for the classifications.}
\format{A tibble of four rows and three columns with the description of the
Bricks used for the classifications.}
\usage{
data(brick_description)
}
\description{
A tibble with the description of the Bricks used for the
classifications.
}
\keyword{datasets}
