% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{castable2numeric}
\alias{castable2numeric}
\title{Test if a character can be casted to numeric.}
\usage{
castable2numeric(chr)
}
\arguments{
\item{chr}{A length-one character.}
}
\value{
A logical.
}
\description{
Test if a character can be casted to numeric.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
