% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compare_ts}
\alias{compare_ts}
\title{Util function for computing similarity parameters between two time series}
\usage{
compare_ts(x, band, sam_l8m_mod)
}
\arguments{
\item{x}{A numeric. A row number in sam_l8m_mod}

\item{band}{A length-one character. The name of a band present in both L8-MOD and MODIS time-series.}

\item{sam_l8m_mod}{A sits tibble with two list-columns for the L8-MOD (ts_l8mod) and MODIS (ts_modis) time series.}
}
\value{
A numeric vector.
}
\description{
Compute comparison parameters between two time series
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
