% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band.R
\name{compose_bands_s2}
\alias{compose_bands_s2}
\title{Build band compositions for the given image's files.}
\usage{
compose_bands_s2(band_tb, out_dir = tempdir())
}
\arguments{
\item{band_tb}{A tibble of files.}

\item{out_dir}{A lenght-one character. Path to a directory for storing the results.}
}
\value{
A tibble.
}
\description{
Build band compositions for the given image's files.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
