% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_indexes}
\alias{compute_indexes}
\title{Compute indexes}
\usage{
compute_indexes(x, sat)
}
\arguments{
\item{x}{A tibble.}

\item{sat}{A Length-one character. The satellite or sensor name in c("Landsat8", "MOD13").}
}
\value{
A tibble with additional variables (prefixed with "c_").
}
\description{
Compute vegetation indexes
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
