% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_sd_raster}
\alias{compute_sd_raster}
\title{Compute the pixel-wise standard deviation.}
\usage{
compute_sd_raster(img, base_dir, cores = 1L)
}
\arguments{
\item{img}{A length-one character. A raster name.}

\item{base_dir}{A length-one character. The path to the rasters.}

\item{cores}{A length-one numeric. The available number of cores.}
}
\description{
Compute the pixel-wise standard deviation of a set of raster
with of same name
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
