% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_sd_raster_gdal}
\alias{compute_sd_raster_gdal}
\title{Compute the pixel-wise standard deviation using GDAL.}
\usage{
compute_sd_raster_gdal(infile, outfile, quiet = TRUE)
}
\arguments{
\item{infile}{The multiband raster file (or a vector of paths to multiple
raster files) for which to calculate cell standard deviations.}

\item{outfile}{Path to raster output file.}

\item{quiet}{Logical. Should gdal_calc.py output be silenced?}
}
\description{
Compute the pixel-wise standard deviation of a set of raster
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
