% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{confusion_raster}
\alias{confusion_raster}
\title{Compute the confusion matrix between two raster files.}
\usage{
confusion_raster(r1_path, r2_path, key_ls)
}
\arguments{
\item{r1_path}{A length-one character. Path to a raster file.}

\item{r2_path}{A length-one character. Path to a raster file.}

\item{key_ls}{A named list used to recode the integer values of r1_path and r2_path. The list is made of labels and its names are the numbers (as character) in the rasters.}
}
\value{
A list as explained in caret::confusionMatrix
export
}
\description{
Compute the confusion matrix between two raster files.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
