% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_get_areas}
\alias{cov_get_areas}
\title{Get the areas}
\usage{
cov_get_areas(cov, label)
}
\arguments{
\item{cov}{A coverage.}

\item{label}{A character. The name of the stratification field.}
}
\value{
A simple feature object of type point (sf).
}
\description{
Get the areas sumarized by label.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
