% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_get_values}
\alias{cov_get_values}
\title{Get values from a coverage}
\usage{
cov_get_values(cov, sample_points)
}
\arguments{
\item{cov}{A coverage.}

\item{sample_points}{A simple feature object of type point (sf).}
}
\value{
A simple feature object of type point (sf).
}
\description{
Get teh values in the coverage at the given points
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
