% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_proj}
\alias{cov_proj}
\title{Transform a coverage's SRS}
\usage{
cov_proj(cov, proj4string)
}
\arguments{
\item{cov}{A coverage.}

\item{proj4string}{A Character. A proj4string description of the target SRS}
}
\value{
An projected object of the same class of cov
}
\description{
Transform a coverage from its spatial reference system to another
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
