% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_read}
\alias{cov_read}
\title{Read a coverage from a file on disk.}
\usage{
cov_read(cov_path)
}
\arguments{
\item{cov_path}{A character. The path to a coverage file.}
}
\value{
A coverage
}
\description{
Read a coverage from a file on disk.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
