% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_sample_random}
\alias{cov_sample_random}
\title{Get random sample points}
\usage{
cov_sample_random(cov, n)
}
\arguments{
\item{cov}{A coverage.}

\item{n}{A numeric. The number of samples.}
}
\value{
A simple feature object (sf) of type point.
}
\description{
Get approximately n samples out of the given coverage
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
