% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_sample_stratified}
\alias{cov_sample_stratified}
\title{Get random sample points}
\usage{
cov_sample_stratified(cov, n, label = NULL, cores = 1L)
}
\arguments{
\item{cov}{A coverage.}

\item{n}{A numeric. The number of samples per class.}

\item{label}{A character. The name of the stratification field.}

\item{cores}{A numeric. The number of cores for parallel processing. The 
defailt is one.}
}
\value{
A simple feature object of type point (sf).
}
\description{
Get approximately n samples out of the given coverage. Each 
label gets approximately the same number of samples
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
