% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_srs}
\alias{cov_srs}
\title{Get a coverage's SRS.}
\usage{
cov_srs(cov)
}
\arguments{
\item{cov}{A coverage.}
}
\value{
A character. The proj4string description of the coverage's SRS
}
\description{
Get a coverage's spatial reference system.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
