% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cov_type}
\alias{cov_type}
\title{Get the type of coverage.}
\usage{
cov_type(cov)
}
\arguments{
\item{cov}{A coverage.}
}
\value{
A character
}
\description{
Test if the given object is either a vector or a raster.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
