% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{deter2mask}
\alias{deter2mask}
\title{Build a mask from DETER.
PRODES's SHP into a TIF.}
\usage{
deter2mask(
  file_path_deter,
  file_path_res,
  pyear,
  out_file = tempfile(pattern = "deter2mask_", fileext = ".tif")
)
}
\arguments{
\item{file_path_deter}{A path to a DETER shapefile.}

\item{file_path_res}{A path to the results of a classification.}

\item{pyear}{A PRODES year.}

\item{out_file}{Optional. Path to an output file.}
}
\value{
A path to a raster file.
}
\description{
Build a raster mask from a DETER shapefile.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
