% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expert_samples}
\alias{expert_samples}
\title{Samples collected by experts.}
\format{An sf object with 39980 rows and 8 colmns including, for each sample,
coordinates, start and end dates, labels, coverage names and time series.}
\usage{
data(expert_samples)
}
\description{
Set of sample points classified by hand by an expert (Ana
Rorato) using MOD13Q1 images.
}
\keyword{datasets}
