% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expert_validation}
\alias{expert_validation}
\title{Validation samples collected by experts.}
\format{An sf object with 450 rows and 9 colmns including Identification,
Landsat tile, and the labels from 2014 to 2017.}
\usage{
data(expert_validation)
}
\description{
Set of sample points for validating classification results. The
samples were selected randomically an classified by hand by an expert
(Rodrigo Begotti) using Landsat mosaics from PRODES Digital.
}
\keyword{datasets}
