% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{f1_score}
\alias{f1_score}
\title{Compute the F1 score.}
\usage{
f1_score(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{A matrix given in sample counts. Columns represent the reference data and rows the results of the classification.}
}
\value{
A tibble with precision, recall and f1 score.
}
\description{
Compute the F1 score from a confusion matrix.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
