% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{get_brick_metadata}
\alias{get_brick_metadata}
\title{Create a tibble with metadata of the bricks in a directory.}
\usage{
get_brick_metadata(brick_dir)
}
\arguments{
\item{brick_dir}{A length-one character. Path to a directory of bricks.}
}
\value{
A tibble.
}
\description{
Create a tibble with metadata of the bricks in a directory.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
