% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{get_classified_maps}
\alias{get_classified_maps}
\title{Get the maps resulting from a classification.}
\usage{
get_classified_maps(dir_path)
}
\arguments{
\item{dir_path}{A length-one character. A path to a directory of tif filles}
}
\value{
A tibble.
}
\description{
Get the maps resulting from a classification.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
