% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_confusion_matrix}
\alias{get_confusion_matrix}
\title{Compute a confusion matrix}
\usage{
get_confusion_matrix(cov_res, cov_ref, n, cores = 1L)
}
\arguments{
\item{cov_res}{A results coverage. It must have a variable "label_res".}

\item{cov_ref}{A reference coverage. It must have a variable "label_ref".}

\item{n}{A numeric. The number of samples.}

\item{cores}{A numeric. The number of cores to use for parallel processing.}
}
\value{
A matrix
}
\description{
Compute a confusion matrix
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
