% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{get_deter_maps}
\alias{get_deter_maps}
\title{Get the DETER maps.}
\usage{
get_deter_maps(dir_path)
}
\arguments{
\item{dir_path}{A length-one character. A path to a directory of shapefilles.}
}
\value{
A tibble.
}
\description{
Get the DETER maps.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
