% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_img_md}
\alias{get_img_md}
\title{Get metadata from a image.}
\usage{
get_img_md(img_path)
}
\arguments{
\item{img_path}{A character. Path to a file.}
}
\value{
A list of character.
export
}
\description{
Get metadata from a image's path.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
