% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{get_prodes_map_key}
\alias{get_prodes_map_key}
\title{Process the classified map labels. Build the key for encoding
PRODES's SHP into a TIF.}
\usage{
get_prodes_map_key(label_file)
}
\arguments{
\item{label_file}{A path to a CSV file with the labels and codes in in the results of the classification.}
}
\value{
A tibble with of PRODES' labels, their reference codes, and their codes in the classification results.
}
\description{
Process the classified map labels.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
