% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{get_quantile_maps}
\alias{get_quantile_maps}
\title{Get the cloud quantile maps.}
\usage{
get_quantile_maps(dir_path)
}
\arguments{
\item{dir_path}{A length-one character. A path to a directory of tif filles}
}
\value{
A tibble.
}
\description{
Get the cloud quantile maps.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
