% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get time series for validates sample points.}
\usage{
get_timeseries(cpath, path_bricks, brick_prefix, class_bands,
  suffix = "", max_time_diff = 30)
}
\arguments{
\item{cpath}{A length-one character. Path to a CSV file.}

\item{path_bricks}{A character. Path to a directory with Landsdat8-MODIS bricks.}

\item{brick_prefix}{A character. Prefix to the name of the Landsdat8-MODIS bricks.}

\item{class_bands}{A character. The name of the bands to process.}

\item{suffix}{A length-one character. A suffix to append to the filename of the results}

\item{max_time_diff}{A length-one numeric. The maximum difference (in days) between the fist day of the bricks and the first day of the CSV sample file.}
}
\value{
A vector of paths to Rdata files
}
\description{
Get validated PRODES samples (from a CSV file) and retrieve time
series for each one
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
