% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_timeseries_modis}
\alias{get_timeseries_modis}
\title{Get MODIS time series for validates sample points.}
\usage{
get_timeseries_modis(cpath)
}
\arguments{
\item{cpath}{A length-one character. Path to a CSV file.}
}
\value{
A length-one character. Path to the result file.
}
\description{
Get validated PRODES samples (from a CSV file) and retrieve time
series for each one from WTSS server.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
