% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{get_water_masks}
\alias{get_water_masks}
\title{Get the water masks.}
\usage{
get_water_masks(dir_path)
}
\arguments{
\item{dir_path}{A length-one character. A path to a directory of tif filles}
}
\value{
A tibble.
}
\description{
Get the water masks.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
