% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{identify_missing_bricks}
\alias{identify_missing_bricks}
\title{Identify bricks missing in directory.}
\usage{
identify_missing_bricks(in_dir, expected_scenes, expected_years,
  expected_bands,
  brick_file_pattern = "^LC8SR-(MASKCLOUD|MOD13Q1-MYD13Q1|MOD13Q1-STARFM|RAW|SIMPLE)_[0-9]{6}_[0-9]{4}-[0-9]{2}-[0-9]{2}_[a-z0-9]{+}_STACK_BRICK.tif$")
}
\arguments{
\item{in_dir}{A length-one character. A path to a directory of brick files.}

\item{expected_scenes}{A character. Expected path and rows.}

\item{expected_years}{A character. Expected years.}

\item{expected_bands}{A character. Expected bands.}

\item{brick_file_pattern}{A length-one character· A regular expression to match brick files.}
}
\value{
A tibble.
}
\description{
Given sets of years, scenes, bands, check if thre are bricks available for them.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
