% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{invert_mask}
\alias{invert_mask}
\title{Swap the logical values in a raster mask.}
\usage{
invert_mask(
  in_file,
  out_file = tempfile(pattern = "invert_mask_", fileext = ".tif")
)
}
\arguments{
\item{in_file}{A length-one character. A path to a raster mask.}

\item{out_file}{A lenght-one character. A path to the file with the results.}
}
\value{
out_file.
}
\description{
Any value greater than one becomes zero and each zero becomes
              one.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
