% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{is_cov}
\alias{is_cov}
\title{Is the given object a coverage?.}
\usage{
is_cov(cov)
}
\arguments{
\item{cov}{A coverage.}
}
\value{
A logical.
}
\description{
Test if the given object is supported.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
