% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_raster_valid}
\alias{is_raster_valid}
\title{Check if gdalinfo is able to open the raster file.}
\usage{
is_raster_valid(file_path)
}
\arguments{
\item{file_path}{A character. Path to a raster file.}
}
\value{
A logical.
}
\description{
Check if gdalinfo is able to open the raster file.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
