% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{load_samples}
\alias{load_samples}
\title{Load sample time series from Rdata files}
\usage{
load_samples(x, sat)
}
\arguments{
\item{x}{A character. A path to a file of samples time series.}

\item{sat}{A Length-one character. The satellite or sensor name in c("Landsat8", "MOD13").}
}
\value{
A tibble or a list of tibbles.
}
\description{
Load (and filter) sample time series from Rdata files.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
