% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapbiomas_labels}
\alias{mapbiomas_labels}
\title{Labels of Mapboimas Amazonia.}
\format{A tibble with 27 rows and 3 variables: Mapbiomas' labels and ids and
        their mapping to PRODES labels.}
\usage{
data(mapbiomas_labels)
}
\description{
Labels for the collection 3 of Mapboimas Amazonia.
}
\keyword{datasets}
