% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mask_raster}
\alias{mask_raster}
\title{Mask a raster.}
\usage{
mask_raster(file_path, mask, band = 1, resampling = "ngb")
}
\arguments{
\item{file_path}{A character. Path to a raster file.}

\item{mask}{Either an sf or raster object. A mask to be applied to the input.}

\item{band}{An integer. The band identifier in the input file.}

\item{resampling}{A character. A resamplig method.}
}
\value{
A raster object.
}
\description{
Mask a raster using either a vector or another raster.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
