% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask_with}
\alias{mask_with}
\title{Apply a mask to a raster.}
\usage{
mask_with(
  in_file,
  mask_file,
  out_file = tempfile(pattern = "mask_with_", fileext = ".tif")
)
}
\arguments{
\item{in_file}{A length-one character. A path to a raster}

\item{mask_file}{A lenght-one character. A path to a raster mask.}

\item{out_file}{A lenght-one character. A path to the file with the results.}
}
\value{
out_file.
}
\description{
Any value greater than one becomes one.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
