% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{parse_training_log}
\alias{parse_training_log}
\title{Parse the log file of a training.}
\usage{
parse_training_log(log_path)
}
\arguments{
\item{log_path}{A length-one character. Path to the log file of a training.}
}
\value{
A list of two: A tibble and a list of tibble
}
\description{
Parse the log file of a training.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
