% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{process_valid_shp}
\alias{process_valid_shp}
\title{Process a shapefile of validated samples.}
\usage{
process_valid_shp(shp_path, out_dir)
}
\arguments{
\item{shp_path}{A length-one character. Path to a shapefile.}

\item{out_dir}{A length-one character. Path to a folder.}
}
\value{
A vector of paths to CSV files
}
\description{
Get validated PRODES samples and split them by label in year
into a table (csv). These processed samples are ready for other script to
retrieve time-series
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
