% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{prodes2mapbiomas}
\alias{prodes2mapbiomas}
\title{Rasterize PRODES to match MAPBIOMAS.}
\usage{
prodes2mapbiomas(file_pd, file_mb, tile, year_mb, prodes_lbl)
}
\arguments{
\item{file_pd}{A length-one character. Path to a PRODES map (a shapefile).}

\item{file_mb}{A length-one character. Path to a MAPBIOMAS map (a tif).}

\item{tile}{A length-one character. The Landsat scene id (6 numbers).}

\item{year_mb}{A length-one integer. The year of the MAPBIOMAS map.}

\item{prodes_lbl}{A tibble mapping the labels of PRODES to MAPBIOMAS. It must contais the PRODES' labels in Portuguese (label_ld_pt, character) and english (label_pd, character), and its ID (id_pd, an integerthat must have a one-to-one relationship to label_id), as well as MAPBIOMAS' label (label_mb, character) and id (id_mb).}
}
\value{
A length-one character. The path to a raster file.
}
\description{
Rasterize a PRODES map to MAPBIOMAS resolution while recoding MAPBIOMAS to PRODES classes.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
