% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{prodes2scene}
\alias{prodes2scene}
\title{Convert the PRODES' scene format to LANDSAT's.}
\usage{
prodes2scene(prodes_scene)
}
\arguments{
\item{prodes_scene}{A character. A PRODES scene id.}
}
\value{
A character. A Landsat scene id.
}
\description{
Convert the PRODES' 5 digit scene Id to Landsat's 6 digit scene Id.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
