% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{prodes_compute_area}
\alias{prodes_compute_area}
\title{Compute PRODES areas.}
\usage{
prodes_compute_area(pd_polygons)
}
\arguments{
\item{pd_polygons}{An sf object (polygons).}
}
\value{
A tibble.
}
\description{
Compute the area of each of the PRODES' labels
}
\examples{
\dontrun{
pd_polygons <- sf::st_read("prodes_233_067.shp", stringsAsFactors = FALSE)
prodes_compute_area(pd_polygons)
}
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
