% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prodes_labels}
\alias{prodes_labels}
\title{Labels of PRODES Amazonia.}
\format{A tibble with 6 rows and 3 variables: PRODES' labels in Portuguesse,}
\usage{
data(prodes_labels)
}
\description{
Labels for the 2017 release of PRODES shapefile.
}
\keyword{datasets}
