% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prodes_mapbiomas}
\alias{prodes_mapbiomas}
\title{Similatity between PRODES and MAPBIOMAS.}
\format{A tibble with 15 rows and 3 variables: A Landsat Tile, a PRODES year,
and a list-colum with the confusion matrix provided by the caret package.}
\usage{
data(prodes_mapbiomas)
}
\description{
Confusion matrices of PRODES (reference map) and MAPBIOMAS. The
comparison was made by mapping the MAPBIOMAS labels to PRODES (see data sets
prodes_labesl and mapbiomas_labels).
}
\keyword{datasets}
