% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{prodes_rasterize}
\alias{prodes_rasterize}
\title{Rasterize PRODES.}
\usage{
prodes_rasterize(ref_path, img_path, pyear, cov_res, level_key_pt,
  level_key, class_name_filter = "FLORESTA")
}
\arguments{
\item{ref_path}{A length-one character. Path to a PRODES shp}

\item{img_path}{A length-one character. Path to a directory}

\item{pyear}{A length-one nueric. A PRODES year}

\item{cov_res}{A results coverage.}

\item{level_key_pt}{A list for recoding values for translating portuguesse to english. See named list on help(recode)}

\item{level_key}{A list for recoding values. See named list on help(recode)}

\item{class_name_filter}{A character. Constrain rasterization to these classes (labels).}
}
\value{
A raster object
}
\description{
Rasterize a PRODES map
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
