% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{raster2mask}
\alias{raster2mask}
\title{Tranform a raster into a mask.}
\usage{
raster2mask(
  in_file,
  out_file = tempfile(pattern = "raster2mask_", fileext = ".tif")
)
}
\arguments{
\item{in_file}{A length-one character. A path to a raster mask.}

\item{out_file}{A lenght-one character. A path to the file with the results.}
}
\value{
out_file.
}
\description{
Any value greater than one becomes one.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
