% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replace_bands_with_random}
\alias{replace_bands_with_random}
\title{Replace column values with random numbers.}
\usage{
replace_bands_with_random(x)
}
\arguments{
\item{x}{A tibble.}
}
\value{
A numeric vector.
}
\description{
Replace the values of the columns (from the second one on) of a tibble uniform random numbers between 0 and 1
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
