% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sample_polygons}
\alias{sample_polygons}
\title{Obtain random samples from polygons}
\usage{
sample_polygons(sf_shape, label, n_samples, min_area, min_dist,
  border_offset, is_density = FALSE)
}
\arguments{
\item{sf_shape}{A POLYGON sf object.}

\item{label}{A length-one character. The name of a variable in sf_shape}

\item{n_samples}{A length-one numeric. The number of samples requested. If is_density = TRUE, n_samples is the number of samples per unit of area.}

\item{min_area}{A length-one numeric. The minimum area of a sampled polygon.}

\item{min_dist}{A length-one numeric. The minimum disatnces between samples.}

\item{border_offset}{A length-one numeric. The minimum distance from the samples to the polygon's borders.}

\item{is_density}{A length-one logical. Is this a density sampling? The dafault is FALSE}
}
\value{
sf_samples   A point sf object with the structure of sf_shape, plues the X and Y coordinates. The number of rows is approximately n_samples.
}
\description{
Get random samples from a polygon sf object. By default, this
function returns approximately the same number of samples for each unique
label. If is_density is true
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
