% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sample_random}
\alias{sample_random}
\title{Get random sample points}
\usage{
sample_random(coverage, n)
}
\arguments{
\item{coverage}{Either a vector or raster coverage.}

\item{n}{Number of samples.}
}
\value{
A simple feature object of type point.
}
\description{
Get approximately n samples out of the given coverage
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
