% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sample_stratified}
\alias{sample_stratified}
\title{Get random samples points}
\usage{
sample_stratified(coverage, n, label, cores = 2L)
}
\arguments{
\item{coverage}{Either a vector or raster coverage.}

\item{n}{Number of samples.}

\item{label}{A length-one character. A label for the samples.}

\item{cores}{A length-one numeric. Number of cores for parallel processing.}
}
\value{
A simple feature object of type point.
}
\description{
Get approximately n samples out of the given coverage. Each label gets approximately the same number of samples
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
