% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_samples.R
\name{sim_sits_tibble}
\alias{sim_sits_tibble}
\title{Simulate a sits_tibble made of random observations}
\usage{
sim_sits_tibble(
  n_samples,
  label = "label_A",
  lon_mean = -65,
  lon_sd = 1,
  lat_mean = -5,
  lat_sd = 1,
  date_start = "2000/01/01",
  date_end = "2000/12/31",
  obs_freq = 23,
  n_vi = 1
)
}
\arguments{
\item{n_samples}{A length-one numeric. The number of samples per label.}

\item{label}{A character. The labels for returned tibble. The default is "label_A".}

\item{lon_mean}{A length-one numeric. The central longitude of the random samples. The default is 65 degrees est.}

\item{lon_sd}{A length-one numeric. The standard deviation of the longitude of the random samples. The default is 1.}

\item{lat_mean}{A length-one numeric. The central latitude of the random samples. The default is 5 degrees south.}

\item{lat_sd}{A length-one numeric. The standard deviation of the latitude of the random samples. The default is 1.}

\item{date_start}{A length-one character.The start date. The default is "2000/01/01".}

\item{date_end}{A length-one character.The start date. The default is "2000/12/31".}

\item{obs_freq}{A length-one numeric. The number of observations in the period. The default are 23.}

\item{n_vi}{A length-one numeric. The number of vegetation indexes on each sample. The default is 1.}
}
\value{
random_st    A sits_tibble
}
\description{
This function simulates a sits_tibble using random data. The
time series are simulated using a cosine function.
}
\examples{
library(sits.prodes)
library(sits)
my_st <- sim_sits_tibble(10, label = c("A", "B"))
sits_plot(my_st)

}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
