% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{simulate_ts}
\alias{simulate_ts}
\title{Simulate a time series of a Vegetation Index (VI). Alber Sanchez}
\usage{
simulate_ts(t_vector, m = 0.5, amplitude = 0.15, freq = 16/365,
  phase_shift = 0, noise_mean = 0, noise_sd = 0.2)
}
\arguments{
\item{t_vector}{Numeric. A vector of time indexes.}

\item{m}{A length-one numeric. The mean of the VI.}

\item{amplitude}{A length-one numeric. The amplitude of the VI wave.}

\item{freq}{A length-one numeric. The frequency of the VI wave.}

\item{phase_shift}{A length-one numeric. The phase shift of the VI wave.}

\item{noise_mean}{A length-one numeric. The mean of the noise.}

\item{noise_sd}{A length-one numeric. The stabndard deviation of the noise.}
}
\value{
A numeric vector
}
\description{
this function simulates a time series of vegetation indexes using a cosine function.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
