% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{splitAt}
\alias{splitAt}
\title{Split a vector at the given positions.}
\usage{
splitAt(x, pos)
}
\arguments{
\item{x}{An atomic vector.}

\item{pos}{A vector os positions}
}
\value{
A list.
}
\description{
Split a vector at the given positions.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
