% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training_logs}
\alias{training_logs}
\title{Logs of the training of Deep Learning.}
\format{A tibble of 14 rows and 4 columns with the experiment, the log's path,
the trainig's setup, and the training's partial results.}
\usage{
data(training_logs)
}
\description{
Logs obtained during the training phase of Deep Learning
classification.
}
\keyword{datasets}
