% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{validate_csv}
\alias{validate_csv}
\title{Validate CSV files of samples.}
\usage{
validate_csv(csv_paths)
}
\arguments{
\item{csv_paths}{A character. Paths to a CSV file.}
}
\value{
A character.
}
\description{
Validate SITS sample CSV files.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
