% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{validate_sampling}
\alias{validate_sampling}
\title{Validate a coverage}
\usage{
validate_sampling(cov_res, cov_ref, lab_res, lab_ref, n_samples = 10000,
  cores = 1)
}
\arguments{
\item{cov_res}{A coverage.}

\item{cov_ref}{A coverage.}

\item{lab_res}{A character. The name of the label field in the results coverage.}

\item{lab_ref}{A character. The name of the label field in the referece coverage.}

\item{n_samples}{A length-one integer. The number of samples to use during validation. The default is 1000.}

\item{cores}{A length-one integer. The number cores to use. The default is 1}
}
\description{
Compare two coverages.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
