% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vector2raster}
\alias{vector2raster}
\title{Rasterize vectors}
\usage{
vector2raster(vector_sf, raster_r, vector_field)
}
\arguments{
\item{vector_sf}{A sf object.}

\item{raster_r}{A raster object. This raster is the reference for coordinate sytem, spatial resolution, and dimensions.}

\item{vector_field}{A length-one character. The name of an integer column in vector_sf.}
}
\value{
A raster layer.
}
\description{
Rasterize a sf object to match the given raster
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
