% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_brick_md}
\alias{get_brick_md}
\title{Get metadata from a Brick filename}
\usage{
get_brick_md(brick_paths)
}
\arguments{
\item{brick_paths}{A character. Path to brick files.}
}
\value{
A character of path, pathrow, start_date, band, year (NA for MODIS).
}
\description{
Get metadata from a Brick's file name.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
