% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get time series of sample points from SITS' bricks.}
\usage{
get_timeseries(
  cpath,
  path_bricks,
  brick_prefix,
  class_bands,
  scale_factor = NULL,
  missing_values = NULL,
  minimum_values = NULL,
  maximum_values = NULL,
  suffix = "",
  max_time_diff = 30,
  cov_name = "SITS coverage",
  time_len = 23,
  time_by = 16
)
}
\arguments{
\item{cpath}{A length-one character. Path to a CSV file.}

\item{path_bricks}{A character. Path to a directory with Landsdat8-MODIS bricks.}

\item{brick_prefix}{A character. Prefix to the name of the Landsdat8-MODIS bricks.}

\item{class_bands}{A character. The name of the bands to process.}

\item{scale_factor}{A list. Scale factor of the image band files.}

\item{missing_values}{A list. Values representing missing pixels for each band.}

\item{minimum_values}{A list. Minimum values of each band.}

\item{maximum_values}{A list. Maximum values of each band.}

\item{suffix}{A length-one character. A suffix to append to the filename of the results}

\item{max_time_diff}{A length-one numeric. The maximum difference (in days) between the fist day of the bricks and the first day of the CSV sample file.}

\item{cov_name}{A length-one character. Name for the sits coverage to create.}

\item{time_len}{A length-one integer. The lenght of the brick's time line.}

\item{time_by}{A length-one inetger. The number of days between observations in the time line.}
}
\value{
A vector of paths to Rdata files
}
\description{
Get validated PRODES samples (from a CSV file) and retrieve time
series for each one
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
