% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{prodes2scene}
\alias{prodes2scene}
\title{Convert the PRODES' scene format to LANDSAT's.}
\usage{
prodes2scene(prodes_scene)
}
\arguments{
\item{prodes_scene}{A length-one character. A PRODES scene id.}
}
\value{
A length-one character. A Landsat scene id.
}
\description{
Insert missing 0s in PRODES' landsat scene id.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
